+++ pypsa/collection.py
@@ -11,6 +11,11 @@ from typing import Any
 
 import pandas as pd
 
+try:
+    from cloudpathlib import AnyPath as Path
+except ImportError:
+    from pathlib import Path
+
 from pypsa._options import options
 from pypsa.definitions.structures import Dict
 from pypsa.networks import Network
@@ -129,17 +134,17 @@ class NetworkCollection:
 
     def __init__(
         self,
-        networks: pd.Series | Sequence[Network | str],
+        networks: pd.Series | Sequence[Network | str | Path],
         index: pd.Index | pd.MultiIndex | Sequence | None = None,
     ) -> None:
         """Initialize the NetworkCollection with one or more networks.
 
         Parameters
         ----------
-        networks : pd.Series | Sequence[Network | str]
-            Sequence or pd.Series of Network objects or strings (file paths/ urls) to
-            include in the collection. If strings are provided, they will be passed to
-            pypsa.Network() to create Network objects.
+        networks : pd.Series | Sequence[Network | str | Path]
+            Sequence or pd.Series of Network objects, strings (file paths/ urls), or
+            Path objects to include in the collection. If strings or Paths are provided,
+            they will be passed to pypsa.Network() to create Network objects.
         index : pd.Index, pd.MultiIndex, Sequence, or None, optional
             The index to use for the collection. If `networks` is of type `pd.Series`,
             no index is allowed and it will be retrieved from the Series. If None, a
@@ -161,10 +166,10 @@ class NetworkCollection:
         def _convert_to_network(item: Any) -> Network:
             if isinstance(item, Network):
                 return item
-            elif isinstance(item, str):
+            elif isinstance(item, (str | Path)):
                 return Network(item)
             else:
-                msg = f"All values must be PyPSA Network objects or strings, got {type(item)}."
+                msg = f"All values must be PyPSA Network objects, strings, or Path objects, got {type(item)}."
                 raise TypeError(msg)
 
         if isinstance(networks, pd.Series):